add tag and permission check

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Comment;
 use App\Http\Requests\CommentFormRequest;
 use App\Http\Requests\PostFormRequest;
 use App\Post;
+use App\Tag;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
@@ -18,7 +19,7 @@ class PostController extends Controller
      */
     public function __construct()
     {
-        $this->middleware('auth', ['except' => ['index', 'hot', 'storeComment', 'show']]);
+        $this->middleware('auth', ['except' => ['index', 'hot', 'storeComment', 'show', 'tag']]);
     }
 
 
@@ -29,7 +30,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::with('user')->orderBy('created_at', 'desc')->paginate(config('siteconfig.per_page'));
+        $posts = Post::with('user', 'tags')->orderBy('created_at', 'desc')->paginate(config('siteconfig.per_page'));
 
         $page_title = '文章總覽';
         return view('posts.index', compact('posts', 'page_title'));
@@ -52,6 +53,14 @@ class PostController extends Controller
         return view('posts.index', compact('posts', 'page_title'));
     }
 
+    public function tag($id)
+    {
+        $tag = Tag::findOrFail($id);
+        $posts = $tag->posts()->with('user','tags')->orderBy('created_at', 'desc')->paginate(config('siteconfig.per_page'));
+        $page_title = '標籤：' . $tag->title;
+        return view('posts.index', compact('posts', 'page_title'));
+    }
+
     /**
      * Show the form for creating a new resource.
      *
@@ -59,8 +68,9 @@ class PostController extends Controller
      */
     public function create()
     {
+        $tags = Tag::all();
         $page_title = '新增文章';
-        return view('posts.create', compact('page_title'));
+        return view('posts.create', compact('page_title', 'tags'));
     }
 
     /**
@@ -71,11 +81,15 @@ class PostController extends Controller
      */
     public function store(PostFormRequest $request)
     {
+        $tags = $request->get('tag');
+
         $post = new Post;
         $post->user_id = 1;
         $post->title = $request->get('title');
         $post->content = $request->get('content');
         $post->save();
+
+        $post->tags()->attach($tags);
 
         return redirect()->route('posts.show', [$post->id])
                         ->with('msg', ['status' => 'success', 'content' => '文章新增完成']);
@@ -104,13 +118,14 @@ class PostController extends Controller
      */
     public function show($id)
     {
-        $post = Post::with('user', 'comments')->findOrFail($id);
+        $post = Post::with('user', 'comments', 'tags')->findOrFail($id);
+        $tags = Tag::all();
         $page_title = $post->title;
 
         $post->page_views++;
         $post->save();
 
-        return view('posts.show', compact('post', 'page_title'));
+        return view('posts.show', compact('post', 'page_title', 'tags'));
     }
 
     /**
@@ -121,10 +136,16 @@ class PostController extends Controller
      */
     public function edit($id)
     {
-        $post = Post::findOrFail($id);
-        $page_title = '編輯文章';
+        $post = Post::with('tags')->findOrFail($id);
 
-        return view('posts.edit', compact('post','page_title'));
+        if($post->is_owner()) {
+            $tags = Tag::all();
+            $page_title = '編輯文章';
+
+            return view( 'posts.edit', compact( 'post', 'page_title', 'tags' ) );
+        }
+
+        return redirect()->route('posts.index');
     }
 
     /**
@@ -137,10 +158,16 @@ class PostController extends Controller
     public function update(PostFormRequest $request, $id)
     {
         $post = Post::findOrFail($id);
-        $post->update($request->all());
 
-        return redirect()->route('posts.show', [$id])
-                         ->with('msg', ['status' => 'success', 'content' => '文章更新完成']);
+        if($post->is_owner()) {
+            $post->update($request->all());
+            $post->tags()->sync($request->get('tag'));
+
+            return redirect()->route('posts.show', [$id])
+                             ->with('msg', ['status' => 'success', 'content' => '文章更新完成']);
+        }
+
+        return redirect()->route('posts.index');
     }
 
     /**
@@ -152,15 +179,20 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        $comment_ids = $post->comments()->lists('id')->toArray();
-        Comment::destroy($comment_ids);
-        $post->delete();
 
-        return redirect()->route('posts.index')
-                         ->with('msg', [
-                             'status' => 'success',
-                             'content' => '文章及回應已刪除。'
-                         ]);
+        if($post->is_owner()) {
+            $comment_ids = $post->comments()->lists('id')->toArray();
+            Comment::destroy($comment_ids);
+            $post->tags()->detach();
+            $post->delete();
 
+            return redirect()->route('posts.index')
+                             ->with('msg', [
+                                 'status' => 'success',
+                                 'content' => '文章及回應已刪除。'
+                             ]);
+        }
+
+        return redirect()->route('posts.index');
     }
 }
